Setup Admin and Oracle Authority (#15)
Reworks the existing authorities to a more generic setup:
*admin*: for config field updates. This can be a multisig.
*oracle_authority*: for updates of history data. This has to be a hot
wallet.

With the previous configuration, the stake_authority had the power to
change itself. Because it had to be a hot wallet, it's much more likely
to get compromised and could then update itself to an attacker's
personal wallet with no recourse from the team. In this configuration,
the admin multisig could set the oracle_authority to a new hot wallet if
the current one has been compromised.