FEATURE: Mark delinquent validators for instant removal (#57)
To cover the edge case when some validators can be removed on the same
epoch, we treat them differently. If there are any validators that are
marked for instant removal - they have to be removed before anything
else can happen.

This could be further cleaned by running epoch maintenance over all
validators. But for now, it should cover us.

---------

Co-authored-by: Christian  Krueger <coach@Christians-MacBook-Pro.local>