Increase CUs for Steward Scoring (#160)
`Issue` -- Steward scoring was stalling. Each `ComputeScore` instruction
consumes roughly 140,000 CUs. And we were packing 10 of them per
transaction. That puts us very close to the 1.4m limit we were
requesting.

`Fix` -- Pack 5 of these instructions per transaction. And bump
requested CU per ix to 200,000 for a cushion.

Also added a new grafana alert that will notify when the number of
scored accounts hasn't progressed in the last hour (and is below the
expected to be scored). Have this alert set to low sev for now. Will
keep an eye on it and tune accordingly.