JIP 25 Impl (#168)
Implementation for JIP-25 

*New steward scoring system of 4-tiers with descending importance*
1) Max inflation commission
2) Avg MEV commission
3) Validator Age (new metric)
4) Avg normalized vote credits

This new score is encoded as a u64 where larger is better, meaning we
can leverage the existing state machine to compute and rebalance
delegations.

*Migration*
The previous / existing score is encoded as a u32 which means we needed
to migrate the `StewardState` account to accommodate the extra bytes. To
ensure the migration is done atomically and preserve the existing state
of the steward state machine, we introduced a versioned account scheme
that updates the account discriminator within the migration instruction.
Extensive test coverage was included to assert the migration copies over
the state byte for byte.

*Validator Age*
A new metric has been introduced into the validator history, that
measures the age of a validator as the number of non-zero vote credit
epochs since program inception. And to account for the scenario where we
may want or need to measure validator age from the cluster genesis epoch
we've included an oracle instruction to upload validator age values
directly.