Add To Blacklist With Squads Multisig (#182)
Adds support for creating a squads proposal directly from the CLI for
adding validators to the steward blacklist.
Also adds support for signing with ledger.

We forked squads-v4 to update the dependencies there to align with what
we have here. The CI started complaining about the new anchor-3 IDL
features, and had to resolve those here as well.
https://github.com/jito-foundation/squads-v4