feat: Directed staking onchain (#185)
# JitoSOL Directed Staking

## Overview
Directed staking enables a permissioned set of stakers to delegate the
SOL underpinning their JitoSOL balance towards specific permissioned
validator preferences. Directed staking delegations will occur at the
start of every epoch and the source-of-truth for directed stake
delegation target amounts will be a permissioned upload operated by Jito
Labs. Certain constraints such as per-epoch un-staking caps and an
undirected stake TVL floor are applied to prevent yield drag and an
imbalance of the stake pool in a way that dis-proportionately favors
directed stake.

## Changes

### Steward State Machine
- **RebalanceDirected** at 0% epoch progress of every epoch  
- **Idle** when `REBALANCE_DIRECTED_COMPLETED` flag is set  
- **ComputeScores** at 50% epoch progress of scoring epoch  
- **ComputeDelegations** when `COMPUTE_SCORES` flag is set during
scoring epoch
- **Idle** when `COMPUTE_DELEGATIONS` set during scoring epoch  
- **ComputeInstantUnstake** at 90% of epoch progress every epoch  
- **Rebalance** when `COMPUTE_INSTANT_UNSTAKE` flag set every epoch  
- **Idle** when StewardState progress is complete every epoch  
- When a new epoch has been detected the `EpochMaintance` will reset
state to `RebalanceDirected` and clear per-epoch flags

## State

### Authorities
- **`directed_stake_whitelist_authority`**  
The permissioned authority for instructions that add or remove
permissioned entities from the whitelist. Practically speaking, this
will be a multi-sig address.

- **`directed_stake_meta_upload_authority`**  
The permissioned authority for instructions that copy target values to
the directed stake meta. Practically speaking, this will be a hot wallet
running on our stakenet keeper instances.

### Parameters
- **`directed_unstake_cap_bps`**  
The maximum percent of stake pool TVL allowed to be unstaked from
directed staking targets in a single epoch.

- **`undirected_stake_floor_lamports`**  
The minimum undirected stake TVL in-order to permit directed stake
rebalance increases.

### Steward State
- **`directed_unstake_total`**  
Tracks the amount of directed stake that has been unstaked in the
current epoch. This is reset when the state-machine transitions out of
the state `RebalanceDirected` to ensure that it starts as zero at the
start of each epoch

## Directed Stake Whitelist
The directed stake whitelist is the source-of-truth for permissioned
stakers and validators. This list differentiates between user directed
stakers and protocol directed stakers for future off-chain meta
calculations and business purposes. Instructions to add and remove
stakers and validators are permissioned by the
`directed_stake_meta_whitelist_authority`.

## Directed Stake Ticket
A directed stake ticket declares a permissioned staker’s delegation
preferences. Delegation preferences must be made up of permissioned
validators and the total number of allocations per-ticket cannot exceed
10,000 basis points.

## Directed Stake Meta
The directed stake meta is the source-of-truth for validator target
delegation amounts and tracking the number of lamports that have been
staked with that validator attributable to directed staking. Each stake
target within the stake meta will track: the vote pubkey of the
validator, the target delegation amount, the staked amount, the epoch
the target was last updated, and the epoch the target was last
rebalanced (and staked amount last updated).

## Instructions

### Directed Stake Whitelist Lifecycle Instructions
- Initialize  
- Realloc  

### Directed Stake Ticket Lifecycle Instructions
- Initialize  
- Close  

### Directed Stake Meta Lifecycle Instructions
- Initialize  
- Realloc  

### Directed Stake Ticket Update Instruction
- `UpdatedDirectedStakeTicket`

### Directed Stake Meta Copy Instruction
- `CopyDirectedStakeTargets`

### Rebalance Instructions
- `Rebalance`  
- `RebalanceDirected`

---------

Co-authored-by: Aoi K <62386689+aoikurokawa@users.noreply.github.com>