steward:  auto transition state when last target is missing from validatorlist (#228)
**Problem**
There is an edge-case where the last validator to directed rebalance has
recently been removed from the validator list. This would result in the
REBALANCE_DIRECTED_COMPLETED flag not being set. This does not result in
the state machine stalling, as the epoch progress conditions would allow
us to transition to Idle later on. However, this is not the exact
behavior we intended.

**Solution**
- Check directed rebalance progress after marking a missing validator as
updated
- maybe_transition after performing the progress check above